agents/pyproject.toml
[project]
name = "team-agents"
version = "0.1.0"
description = "AI Multi-Agent Team for Web Service Development"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
